refactor: move Tooltip component to common directory for better organization
